implement create_task_template.json and update_task_template.json

+++ rill-flow-dag/olympicene-storage/src/main/java/com/weibo/rill/flow/olympicene/storage/dao/mapper/TaskTemplateDAO.java
@@ -29,7 +29,7 @@ import java.util.List;
 @Component
 public interface TaskTemplateDAO {
     @InsertProvider(type = TaskTemplateProvider.class, method = "insert")
-    int insert(TaskTemplateDO taskTemplateDO);
+    long insert(TaskTemplateDO taskTemplateDO);
 
     @SelectProvider(type = TaskTemplateProvider.class, method = "getTaskTemplateList")
     List<TaskTemplateDO> getTaskTemplateList(TaskTemplateParams params);
